Implemented the third major revision of the LTL model checker. This time I reshuffled the application of the different analysis steps such that the entire LTL formula is evaluated every time when visting a state in the fixpoint algorithm. This results in a major speedup and "improved correctness". make check reports that verifier1 & 2 disagree in a number of cases, however counterexamples and the new verifier2 do agree.

+++ projects/CodeThorn/src/LTL.h
@@ -349,11 +349,11 @@ namespace LTL {
 
     operator std::string () const { return std::string(e); }
     operator const Expr& () const { return e; }
-    short size() const { return expr_size; }
+    size_t size() const { return expr_size; }
 
   protected:
     Expr& e;
-    short expr_size;
+    size_t expr_size;
   };
 
 } // namespace LTL
